+++ internal/models/user/user.go
@@ -4,7 +4,7 @@ import "time"
 
 // User represents a user in the database
 type User struct {
-	ID           int       `db:"id" json:"id"`
+	ID           int64     `db:"id" json:"id"`
 	FullName     string    `db:"full_name" json:"full_name"`
 	Email        string    `db:"email" json:"email"`
 	PasswordHash string    `db:"password_hash" json:"-"` // Never expose password hash in JSON
